requirements.txt
numpy>=1.22.0
torch>=1.12.0
torchvision>=0.13.0
opencv-python>=4.5.0
trimesh>=3.9.0
pillow>=9.0.0
matplotlib>=3.5.0
diffusers>=0.12.0
transformers>=4.21.0
pyrender>=0.1.45
scikit-image>=0.19.0
scipy>=1.8.0